#30 Kontrollerek módosítsa az új modelekhezz megfelelően
