+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/ComplexNumberRecordRequestDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
