+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/EmployeeDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.Builder;
 import lombok.Data;
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class EmployeeDto {
-    private int employeeID;
+    private int id;
     private String employeeName;
     private String employeeEmail;
     private String pwd;
