+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/EmployeeRecordRequestDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class EmployeeRecordRequestDto {
-    private int employeeID;
+    private int id;
     private String employeeName;
     private String employeeEmail;
     private String pwd;
