+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/GuestDto.java
@@ -1,20 +1,18 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import java.util.Date;
 
+
 @Data
 @Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class GuestRecordRequestDto {
-    private String name;
+public class GuestDto {
     private String identityNumber;
+    private String name;
     private Date dateOfBirth;
     private Date arrivalDate;
     private Date departureDate;
+
 }
