+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/GuestRecordRequestDto.java
@@ -1,19 +1,20 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.util.Date;
-
 
 @Data
 @Builder
-public class GuestDto {
-    private String name;
+@NoArgsConstructor
+@AllArgsConstructor
+public class GuestRecordRequestDto {
     private String identityNumber;
+    private String name;
     private Date dateOfBirth;
     private Date arrivalDate;
     private Date departureDate;
-
 }
