+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/PositionRecordRequestDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 
 import lombok.Builder;
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 @Builder
-public class PositionDto {
+public class PositionRecordRequestDto {
     private String id;
     private String name;
 }
