+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/ProductsDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.Builder;
 import lombok.Data;
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 @Builder
 public class ProductsDto{
+        private int id;
         private String productName;
         private int productPrice;
-        private int barCode;
-        private boolean Adult;
+        private int Adult;
 }
