+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/ServiceDto.java
@@ -1,11 +1,11 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class ServiceRecordRequestDto {
+public class ServiceDto {
     private String id;
     private String name;
     private int price;
