+++ backend/src/main/java/hu/uni/eku/team3/controller/dto/SummaryDto.java
@@ -1,4 +1,4 @@
-package hu.uni.eku.tzs.controller.dto;
+package hu.uni.eku.team3.controller.dto;
 
 import lombok.Builder;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 
 public class SummaryDto {
-    private int summaryID;
+    private int id;
     private int identityNumber;
     private int servicesID;
     private int barCode;
