--- backend/src/main/java/hu/uni/eku/tzs/controller/dto/SummaryRecordRequestDto.java
@@ -1,18 +1,0 @@
-package hu.uni.eku.tzs.controller.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SummaryRecordRequestDto{
-    private int summaryID;
-    private int identityNumber;
-    private int servicesID;
-    private int barCode;
-}
-
